Add Firebase encoder
